Move keys_stringifier into jsonapi_errors_handler
