+++ lib/jsonapi_errors_handler/keys_stringifier.rb
@@ -1,0 +1,10 @@
+module JsonapiErrorsHandler
+  class KeysStringifier
+    def self.call(hash)
+      hash.reduce({}) do |h, (k,v)|
+        new_val = v.is_a?(Hash) ? self.call(v) : v
+        h.merge({k.to_s => new_val})
+      end
+    end
+  end
+end
